Stop importing from __init__ file of own pkg
Split declarations into smaller files
